requirements.txt
numpy
pandas
opencv-python
pyttsx3
vosk
openai-whisper
sounddevice
soundfile
torch
torchvision
transformers
sentencepiece
llama-cpp-python
flask
flask-socketio
python-socketio
requests
adb-shell
pure-python-adb
pycryptodome
pyautogui
pynput
speechbrain
beautifulsoup4
selenium
